requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
arabic-reshaper==3.0.0
asn1crypto==1.5.1
asttokens==3.0.1
attrs==25.4.0
blinker==1.9.0
boto3==1.42.78
botocore==1.42.78
brotli==1.2.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
choreographer==1.2.1
click==8.3.1
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.5
cssselect2==0.9.0
cycler==0.12.1
debugpy==1.8.20
decorator==5.2.1
executing==2.2.1
fastapi==0.129.0
fastjsonschema==2.21.2
Flask==3.1.3
fonttools==4.61.1
freetype-py==2.5.1
greenlet==3.3.2
h11==0.16.0
html5lib==1.1
idna==3.11
iniconfig==2.3.0
ipykernel==7.2.0
ipython==9.11.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.1.0
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
kaleido==1.2.0
kiwisolver==1.4.9
lightgbm==4.6.0
logistro==2.0.1
lxml==6.0.2
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
narwhals==2.16.0
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==2.4.2
orjson==3.11.7
oscrypto==1.3.0
packaging==26.0
pandas==3.0.1
parso==0.8.6
pillow==12.1.1
platformdirs==4.9.4
playwright==1.58.0
plotly==6.5.2
pluggy==1.6.0
prompt_toolkit==3.0.52
psutil==7.2.2
pure_eval==0.2.3
pycairo==1.29.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydyf==0.12.1
pyee==13.0.1
Pygments==2.19.2
pyHanko==0.33.0
pyhanko-certvalidator==0.29.1
pyparsing==3.3.2
pypdf==6.7.4
pyphen==0.17.2
pytest==9.0.2
pytest-timeout==2.4.0
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
reportlab==4.4.10
requests==2.32.5
resend==2.25.0
rlPyCairo==0.4.0
rpds-py==0.30.0
s3transfer==0.16.0
scikit-learn==1.8.0
scipy==1.17.0
simplejson==3.20.2
six==1.17.0
stack-data==0.6.3
starlette==0.52.1
svglib==1.6.0
threadpoolctl==3.6.0
tinycss2==1.5.1
tinyhtml5==2.1.0
tornado==6.5.5
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
Unidecode==1.4.0
uritools==6.0.1
urllib3==2.6.3
uvicorn==0.41.0
waitress==3.0.2
wcwidth==0.6.0
weasyprint==68.1
webencodings==0.5.1
Werkzeug==3.1.6
xhtml2pdf==0.2.17
zopfli==0.4.1
